+++ src/main/java/com/capstone/BnagFer/domain/accounts/dto/social/UserSocialLoginRequestDto.java
@@ -1,0 +1,6 @@
+package com.capstone.BnagFer.domain.accounts.dto.social;
+
+public record UserSocialLoginRequestDto(
+        String accessToken
+) {
+}
